// shell/ui/src/generated/Event.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { App } from "./App";
import type { ServingState } from "./ServingState";
import type { Viewer } from "./Viewer";

export type Event = { "event": "app_changed", app: App, } | { "event": "app_removed", slug: string, } | { "event": "serving_changed", serving: ServingState, } | { "event": "pairing_request", device_id: string, } | { "event": "access", app_slug?: string, action: string, } | { "event": "presence", app_slug: string, viewers: Array<Viewer>, } | { "event": "error", message: string, };